requirements.txt
# Orcabikes Telegram Bot Project: Python libraries we'll be using for this project.

# Telegram Bot
python-telegram-bot
requests

# For DynamoDB Database
boto3==1.20.23
botocore==1.23.23